Fix initial solution injection  - Add solution update  - Remove wrong vars-vals

+++ core/src/core/data/attributes/Consumer.hpp
@@ -7,7 +7,7 @@
 #include "core/data/attributes.hpp"
 
 namespace routing {
-    typedef double Demand;
+    typedef int Demand;
     namespace attributes {
 
         /**
